chore: move test-utils to the top level of packages
