--- packages/core/medusa-test-utils/tsconfig.json
@@ -1,3 +1,0 @@
-{
-  "extends": "../../../_tsconfig.base.json"
-}
