+++ packages/medusa-test-utils/src/index.ts
@@ -4,6 +4,6 @@ export { default as IdMap } from "./id-map"
 export * from "./init-modules"
 export * as JestUtils from "./jest"
 export * from "./medusa-test-runner"
-export * from "./medusa-test-runner-utils/"
+export * from "./medusa-test-runner-utils"
 export { default as MockEventBusService } from "./mock-event-bus-service"
 export * from "./module-test-runner"
