+++ packages/medusa-test-utils/src/medusa-test-runner-utils/use-db.ts
@@ -54,7 +54,6 @@ export async function syncLinks(
 
 async function loadCustomLinks(directory: string, container: MedusaContainer) {
   // TODO: move to framework once settle down
-  await import("@medusajs/medusa/loaders/index")
   const { getResolvedPlugins } = await import(
     "@medusajs/medusa/loaders/helpers/resolve-plugins"
   )
